open config json messaging mixup
