+++ core/core.ts
@@ -28,7 +28,11 @@ import { DevDataSqliteDb } from "./util/devdataSqlite";
 import { fetchwithRequestOptions } from "./util/fetchWithOptions";
 import { GlobalContext } from "./util/GlobalContext";
 import historyManager from "./util/history";
-import { editConfigJson, setupInitialDotContinueDirectory } from "./util/paths";
+import {
+  editConfigJson,
+  getConfigJsonPath,
+  setupInitialDotContinueDirectory,
+} from "./util/paths";
 import { Telemetry } from "./util/posthog";
 import { TTS } from "./util/tts";
 
@@ -256,6 +260,10 @@ export class Core {
         (await this.config()).experimental?.promptPath,
       );
       await this.configHandler.reloadConfig();
+    });
+
+    on("config/openLocalConfigFile", async (msg) => {
+      await this.ide.openFile(getConfigJsonPath());
     });
 
     on("config/reload", (msg) => {
