+++ core/protocol/core.ts
@@ -67,6 +67,7 @@ export type ToCoreFromIdeOrWebviewProtocol = {
   "config/deleteModel": [{ title: string }, void];
   "config/reload": [undefined, BrowserSerializedContinueConfig];
   "config/listProfiles": [undefined, ProfileDescription[]];
+  "config/openLocalConfigFile": [undefined, void];
   "context/getContextItems": [
     {
       name: string;
