+++ core/protocol/ideWebview.ts
@@ -29,7 +29,6 @@ export type ToIdeFromWebviewProtocol = ToIdeFromWebviewOrCoreProtocol & {
   ];
   showTutorial: [undefined, void];
   showFile: [{ filepath: string }, void];
-  openConfig: [undefined, void];
   toggleDevTools: [undefined, void];
   reloadWindow: [undefined, void];
   focusEditor: [undefined, void];
@@ -94,7 +93,6 @@ export type ToWebviewFromIdeProtocol = ToWebviewFromIdeOrCoreProtocol & {
   navigateTo: [{ path: string; toggle?: boolean }, void];
   addModel: [undefined, void];
 
-  openSettings: [undefined, void];
   /**
    * @deprecated Use navigateTo with a path instead.
    */
