+++ core/protocol/webview.ts
@@ -5,6 +5,7 @@ import type { ContextItemWithId, IndexingProgressUpdate } from "../index.js";
 export type ToWebviewFromIdeOrCoreProtocol = {
   configUpdate: [undefined, void];
   configError: [ConfigValidationError[] | undefined, void];
+  openSelectedConfigProfile: [undefined, void];
   getDefaultModelTitle: [undefined, string];
   indexProgress: [IndexingProgressUpdate, void];
   refreshSubmenuItems: [undefined, void];
@@ -18,4 +19,6 @@ export type ToWebviewFromIdeOrCoreProtocol = {
   ];
   setTTSActive: [boolean, void];
   getWebviewHistoryLength: [undefined, number];
+  signInToControlPlane: [undefined, void];
+  openDialogMessage: ["profileSwitcher", void];
 };
