+++ extensions/intellij/src/main/kotlin/com/github/continuedev/continueintellijextension/toolWindow/ContinueBrowser.kt
@@ -192,9 +192,6 @@ class ContinueBrowser(val project: Project, url: String) {
                 }
 
                 "reloadWindow" -> {}
-                "openConfig" -> {
-                    ide?.setFileOpen(getConfigJsonPath())
-                }
 
                 "readRangeInFile" -> {
                     val data = data.asJsonObject
