+++ extensions/vscode/src/extension/VsCodeExtension.ts
@@ -317,9 +317,13 @@ export class VsCodeExtension {
 
     // When GitHub sign-in status changes, reload config
     vscode.authentication.onDidChangeSessions(async (e) => {
-      if (e.provider.id === "github") {
-        this.configHandler.reloadConfig();
-      } else if (e.provider.id === controlPlaneEnv.AUTH_TYPE) {
+      if (e.provider.id === controlPlaneEnv.AUTH_TYPE) {
+        vscode.commands.executeCommand(
+          "setContext",
+          "continue.isSignedInToControlPlane",
+          true,
+        );
+
         const sessionInfo = await getControlPlaneSessionInfo(true);
         this.webviewProtocolPromise.then(async (webviewProtocol) => {
           void webviewProtocol.request("didChangeControlPlaneSessionInfo", {
@@ -329,7 +333,19 @@ export class VsCodeExtension {
           // To make sure continue-proxy models and anything else requiring it get updated access token
           this.configHandler.reloadConfig();
         });
-        void this.core.invoke("didChangeControlPlaneSessionInfo", { sessionInfo });
+        void this.core.invoke("didChangeControlPlaneSessionInfo", {
+          sessionInfo,
+        });
+      } else {
+        vscode.commands.executeCommand(
+          "setContext",
+          "continue.isSignedInToControlPlane",
+          false,
+        );
+
+        if (e.provider.id === "github") {
+          this.configHandler.reloadConfig();
+        }
       }
     });
