+++ gui/src/components/AddModelButtonSubtext.tsx
@@ -10,7 +10,9 @@ function AddModelButtonSubtext() {
       This will update your{" "}
       <span
         className="cursor-pointer underline"
-        onClick={() => ideMessenger.post("openConfig", undefined)}
+        onClick={() =>
+          ideMessenger.post("config/openLocalConfigFile", undefined)
+        }
       >
         config file
       </span>
