+++ gui/src/components/Layout.tsx
@@ -12,7 +12,11 @@ import {
   setEditStatus,
   startEditMode,
 } from "../redux/slices/editModeState";
-import { setShowDialog, updateApplyState } from "../redux/slices/uiStateSlice";
+import {
+  setDialogMessage,
+  setShowDialog,
+  updateApplyState,
+} from "../redux/slices/uiStateSlice";
 import { RootState } from "../redux/store";
 import { getFontSize, isMetaEquivalentKeyPressed } from "../util";
 import { getLocalStorage, setLocalStorage } from "../util/localStorage";
@@ -21,6 +25,8 @@ import TextDialog from "./dialogs";
 import Footer from "./Footer";
 import { isNewUserOnboarding, useOnboardingCard } from "./OnboardingCard";
 import PostHogPageView from "./PosthogPageView";
+import ProfileSwitcherDialog from "./ProfileSwitcherDialog";
+import { AuthProvider } from "../context/Auth";
 
 const LayoutTopDiv = styled(CustomScrollbarDiv)`
   height: 100%;
@@ -106,16 +112,23 @@ const Layout = () => {
   }, [timeline]);
 
   useWebviewListener(
+    "openDialogMessage",
+    async (message) => {
+      if (message === "profileSwitcher") {
+        dispatch(setShowDialog(true));
+        dispatch(setDialogMessage(<ProfileSwitcherDialog />));
+      }
+    },
+    [],
+  );
+
+  useWebviewListener(
     "addModel",
     async () => {
       navigate("/models");
     },
     [navigate],
   );
-
-  useWebviewListener("openSettings", async () => {
-    ideMessenger.post("openConfig", undefined);
-  });
 
   useWebviewListener(
     "viewHistory",
@@ -207,56 +220,58 @@ const Layout = () => {
   }, [location]);
 
   return (
-    <LastSessionProvider>
-      <LayoutTopDiv>
-        <div
-          style={{
-            scrollbarGutter: "stable both-edges",
-            minHeight: "100%",
-            display: "grid",
-            gridTemplateRows: "1fr auto",
-          }}
-        >
-          <TextDialog
-            showDialog={showDialog}
-            onEnter={() => {
-              dispatch(setShowDialog(false));
+    <AuthProvider>
+      <LastSessionProvider>
+        <LayoutTopDiv>
+          <div
+            style={{
+              scrollbarGutter: "stable both-edges",
+              minHeight: "100%",
+              display: "grid",
+              gridTemplateRows: "1fr auto",
             }}
-            onClose={() => {
-              dispatch(setShowDialog(false));
-            }}
-            message={dialogMessage}
+          >
+            <TextDialog
+              showDialog={showDialog}
+              onEnter={() => {
+                dispatch(setShowDialog(false));
+              }}
+              onClose={() => {
+                dispatch(setShowDialog(false));
+              }}
+              message={dialogMessage}
+            />
+
+            <GridDiv className="">
+              <PostHogPageView />
+              <Outlet />
+
+              {hasFatalErrors && pathname !== ROUTES.CONFIG_ERROR && (
+                <div
+                  className="z-50 cursor-pointer bg-red-600 p-4 text-center text-white"
+                  role="alert"
+                  onClick={() => navigate(ROUTES.CONFIG_ERROR)}
+                >
+                  <strong className="font-bold">Error!</strong>{" "}
+                  <span className="block sm:inline">
+                    Could not load config.json
+                  </span>
+                  <div className="mt-2 underline">Learn More</div>
+                </div>
+              )}
+
+              <ModelDropdownPortalDiv id="model-select-top-div"></ModelDropdownPortalDiv>
+              <ProfileDropdownPortalDiv id="profile-select-top-div"></ProfileDropdownPortalDiv>
+              <Footer />
+            </GridDiv>
+          </div>
+          <div
+            style={{ fontSize: `${getFontSize() - 4}px` }}
+            id="tooltip-portal-div"
           />
-
-          <GridDiv className="">
-            <PostHogPageView />
-            <Outlet />
-
-            {hasFatalErrors && pathname !== ROUTES.CONFIG_ERROR && (
-              <div
-                className="z-50 cursor-pointer bg-red-600 p-4 text-center text-white"
-                role="alert"
-                onClick={() => navigate(ROUTES.CONFIG_ERROR)}
-              >
-                <strong className="font-bold">Error!</strong>{" "}
-                <span className="block sm:inline">
-                  Could not load config.json
-                </span>
-                <div className="mt-2 underline">Learn More</div>
-              </div>
-            )}
-
-            <ModelDropdownPortalDiv id="model-select-top-div"></ModelDropdownPortalDiv>
-            <ProfileDropdownPortalDiv id="profile-select-top-div"></ProfileDropdownPortalDiv>
-            <Footer />
-          </GridDiv>
-        </div>
-        <div
-          style={{ fontSize: `${getFontSize() - 4}px` }}
-          id="tooltip-portal-div"
-        />
-      </LayoutTopDiv>
-    </LastSessionProvider>
+        </LayoutTopDiv>
+      </LastSessionProvider>
+    </AuthProvider>
   );
 };
